load

+++ db/compaction/compaction.cc
@@ -214,7 +214,8 @@ Compaction::Compaction(VersionStorageInfo* vstorage,
                        std::vector<FileMetaData*> _grandparents,
                        bool _manual_compaction, double _score,
                        bool _deletion_compaction,
-                       CompactionReason _compaction_reason)
+                       CompactionReason _compaction_reason,
+                       ColumnCompactionItem *ccitem)
     : input_vstorage_(vstorage),
       start_level_(_inputs[0].level),
       output_level_(_output_level),
@@ -237,7 +238,8 @@ Compaction::Compaction(VersionStorageInfo* vstorage,
       is_full_compaction_(IsFullCompaction(vstorage, inputs_)),
       is_manual_compaction_(_manual_compaction),
       is_trivial_move_(false),
-      compaction_reason_(_compaction_reason) {
+      compaction_reason_(_compaction_reason),
+      ccitem_(ccitem){
   MarkFilesBeingCompacted(true);
   if (is_manual_compaction_) {
     compaction_reason_ = CompactionReason::kManualCompaction;
@@ -271,6 +273,7 @@ Compaction::Compaction(VersionStorageInfo* vstorage,
 }
 
 Compaction::~Compaction() {
+  if (ccitem_) delete ccitem_;
   if (input_version_ != nullptr) {
     input_version_->Unref();
   }
@@ -300,6 +303,11 @@ bool Compaction::IsTrivialMove() const {
   // a very expensive merge later on.
   // If start_level_== output_level_, the purpose is to force compaction
   // filter to be applied to that level, and thus cannot be a trivial move.
+///
+  if (ccitem_ !=nullptr && start_level_ == 0){
+    return false;
+  }
+///
 
   // Check if start level have files with overlapping ranges
   if (start_level_ == 0 && input_vstorage_->level0_non_overlapping() == false) {
@@ -349,11 +357,52 @@ bool Compaction::IsTrivialMove() const {
 
 void Compaction::AddInputDeletions(VersionEdit* out_edit) {
   for (size_t which = 0; which < num_input_levels(); which++) {
+    if(ccitem_ != nullptr && level(which) == 0) continue;
     for (size_t i = 0; i < inputs_[which].size(); i++) {
       out_edit->DeleteFile(level(which), inputs_[which][i]->fd.GetNumber());
     }
   }
+  ///
+  if(ccitem_ == nullptr) return;
+  FileEntry* file = nullptr;
+  FileMetaData* filemeta = nullptr;
+  for(unsigned int i = 0;i < ccitem_->files.size();i++){
+    file = ccitem_->files.at(i);
+    filemeta = ccitem_->L0compactionfiles.at(i);
+    if(filemeta->first_key_index + ccitem_->keys_num.at(i) == file->keys_num){ //该文件已compaction完
+      out_edit->DeleteFile(0,file->filenum);
+      column_compaction_delete_file_.push_back(file->filenum);
+    }
+    else{  //该文件还有数据
+      out_edit->DeleteFile(0,file->filenum);
+      uint64_t file_size = filemeta->fd.GetFileSize() - ccitem_->keys_size.at(i);
+      uint64_t first_key_index = filemeta->first_key_index + ccitem_->keys_num.at(i);
+      InternalKey smallest = file->keys_meta[first_key_index].key;
+      //TODO:smallest_seqno and largest_seqno need update
+      out_edit->AddFile(0 /* level */, filemeta->fd.GetNumber(), filemeta->fd.GetPathId(),
+                   file_size, smallest, filemeta->largest,
+                   filemeta->fd.smallest_seqno, filemeta->fd.largest_seqno,
+                   filemeta->marked_for_compaction, filemeta->is_nvm_level0, first_key_index,filemeta->nvm_sstable_index,
+                   filemeta->keys_num, filemeta->key_point_filenum, filemeta->raw_file_size, filemeta->nvm_meta_size);
+    }
+  }
+  ///
 }
+
+///
+void Compaction::InstallColumnCompactionItem(){
+  if(ccitem_ == nullptr) return;
+  cfd_->set_bg_column_compaction(false);
+
+  RECORD_LOG("InstallColumnCompactionItem delete:[");
+  for(unsigned int i = 0;i < column_compaction_delete_file_.size(); i++){
+    cfd_->nvmcfmodule->DeleteColumnCompactionFile(column_compaction_delete_file_[i]);
+    RECORD_LOG("%ld ",column_compaction_delete_file_[i]);
+  }
+  RECORD_LOG("]\n");
+
+}
+////
 
 bool Compaction::KeyNotExistsBeyondOutputLevel(
     const Slice& user_key, std::vector<size_t>* level_ptrs) const {
